+++ src/app/types/api.ts
@@ -1,0 +1,9 @@
+export interface Pagination {
+  next_key: string;
+}
+
+export interface QueryMeta {
+  next: () => void;
+  hasMore: boolean;
+  isFetchingMore: boolean;
+}
